requirements.txt
PyCryptodome
loguru